iotdevices/requirements.txt
aiocoap==0.4.7
psycopg2-binary==2.9.9
python-dateutil==2.8.2
PyYAML==6.0.1
